Fix: server socket 및 ssh 연결 시점을 변경
socket이 client와 connection을 할때 ssh를 연결하고 socket과 ssh간에
데이터를 흘려보낼 수 있도록 변경
