+++ server/src/utils/ssh-manager.js
@@ -7,6 +7,9 @@ class SshConnectionManager {
 
   async makeShellConnection(session) {
     const { id } = session;
+    if (this.connections[id]) {
+      return this.connections[id].channel;
+    }
     this.connections[id] = new SshConnection();
     const current = this.connections[id];
     const shellChannel = await current.connect({ ...session });
@@ -18,4 +21,4 @@ class SshConnectionManager {
   }
 }
 
-export default new SshConnectionManager();
+module.exports = new SshConnectionManager();
